minor refactor + remove VR images
