+++ requirements.txt
@@ -1,4 +1,3 @@
 numpy>=1.19.0
 opencv-python>=4.2.0.34
 matplotlib>=3.2.2
-imutils>=0.5.3
